Update wget.asm
Add support for second processor.
If present data transfers will go to the second processor unless flags are set. The load address will default to PAGE (&800) unless supplied.
A new flag F will force loading into the host processor. Any other flags will do the same.
If the load address is >= &8000 it is a ROM and a run command will be sent to the second processor. Care must be taken that the correct load address is supplied, ie. usually &8000 or &B800. If it is wrong the system will probably crash!