Firmware version 1.3.0.2 added
Just in case you break your ESP module or get a module with an incompatible or broken firmware. Use the flash_esp.sh script on Linux to program the module. Requires esptool.py and a suitable programmer board. It's not possible to program the ESP module on the WiFi board.